+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Space key toggles FK preview popover on selected cell, rebindable in Settings > Keyboard (#648)
+
 ## [0.29.0] - 2026-04-09
 
 ### Added